refactor: implement 4-layer clean architecture with infrastructure layer
- Add infrastructure layer for NestJS module wiring and DI
- Move event handlers from domain to application layer
- Simplify useCase structure (flatten folders, merge DTOs into ports)
- Fix layer boundary violations (ports no longer import from adapters)
- Add @ianvs/prettier-plugin-sort-imports for consistent import ordering
- Update CLAUDE.md and TODO.md with new architecture documentation

BREAKING CHANGE: Module imports changed from CreateOrderModule to OrderInfrastructureModule
