+++ .prettierrc
@@ -4,5 +4,16 @@
   "semi": true,
   "singleQuote": true,
   "printWidth": 120,
-  "useTabs": false
+  "useTabs": false,
+  "plugins": ["@ianvs/prettier-plugin-sort-imports"],
+  "importOrder": [
+    "^@nestjs/(.*)$",
+    "",
+    "^@config/(.*)$",
+    "^@shared/(.*)$",
+    "",
+    "^@modules/(.*)$",
+    "",
+    "^[./]"
+  ]
 }
