+++ package.json
@@ -61,6 +61,7 @@
   },
   "devDependencies": {
     "@hewmen/serverless-plugin-typescript": "^1.1.17",
+    "@ianvs/prettier-plugin-sort-imports": "^4.7.0",
     "@nestjs/cli": "^10.3.2",
     "@nestjs/schematics": "^10.1.1",
     "@nestjs/testing": "^10.3.8",
