+++ src/config/providers/database.config.ts
@@ -1,5 +1,6 @@
 import { readFile } from 'fs/promises';
 import { parse } from 'dotenv';
+
 import { EnvConfig, IDatabaseConfig } from '@config/config.types';
 
 export class DataBaseConfig implements IDatabaseConfig {
