+++ src/main.ts
@@ -1,10 +1,8 @@
 // ORDER MODULE
 import './modules/order/application/ms/http';
 import './modules/order/application/ms/websocket';
-
 // PRODUCT MODULE
 import './modules/products/application/ms/http';
-
 // LOGGER MODULE
 import './modules/logger/application/ms/http';
 import './modules/logger/application/ms/tcp';
