+++ src/modules/logger/application/ms/http/http.module.spec.ts
@@ -1,6 +1,7 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { INestApplication } from '@nestjs/common';
 import request from 'supertest';
+import { INestApplication } from '@nestjs/common';
+import { Test, TestingModule } from '@nestjs/testing';
+
 import { HttpModule } from './http.module';
 
 describe('AppController (e2e)', () => {
