+++ src/modules/order/application/ms/http/api/api.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Post, Inject, LoggerService, Body } from '@nestjs/common';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { ApiTags } from '@nestjs/swagger';
 
-import { CreateOrderUseCaseResponse } from '@modules/order/application/useCases/CreateOrder/CreateOrder.dto';
+import { CreateOrderUseCaseResponse } from '@modules/products/application/useCases/CreateProduct/CreateProduct.dto';
 
 import { CreateOrderDTO } from './api.dto';
 import ClientsService from './api.service';
