+++ src/modules/order/application/ms/websocket/index.ts
@@ -1,4 +1,5 @@
 import { NestFactory } from '@nestjs/core';
+
 import { WebsocketGatewayModule } from './websocket.module';
 
 async function bootstrap() {
