+++ src/modules/order/application/ports/orderService.port.ts
@@ -1,8 +1,9 @@
-import { Order as OrderSchema } from '../../adapters/repository/order.schema';
 import { Order as OrderEntity } from '../../domain/order';
 
+export interface CreateOrderDTO {
+  items: string[];
+}
+
 export interface OrderServicePort {
-  getAllOrders(): Promise<OrderSchema[]>;
-  getOrderById(id: string): Promise<OrderSchema>;
-  createOrder(order: OrderSchema, entity: OrderEntity): Promise<OrderSchema>;
+  createOrder(order: CreateOrderDTO, entity: OrderEntity): Promise<CreateOrderDTO>;
 }
