+++ src/modules/products/adapters/repository/order.interface.ts
@@ -1,4 +1,5 @@
 import { Repository } from '@shared/adapters/repository/interface';
+
 import { OrderProps } from '@modules/order/domain/order';
 
 export interface DataService {
