+++ src/modules/products/application/ms/http/api/api.service.ts
@@ -1,10 +1,7 @@
 import { Injectable } from '@nestjs/common';
-
 import { CreateOrderUseCase } from '@modules/products/application/useCases/CreateProduct/CreateProduct.usecase';
-import {
-  CreateOrderDTO,
-  CreateOrderUseCaseResponse,
-} from '@modules/order/application/useCases/CreateOrder/CreateOrder.dto';
+import { CreateOrderDTO } from '@modules/order/application/ports/orderService.port';
+import { CreateOrderUseCaseResponse } from '@modules/products/application/useCases/CreateProduct/CreateProduct.dto';
 
 @Injectable()
 export default class ClientsService {
