+++ src/modules/products/domain/events/emitters/orderCreated.emitter.ts
@@ -1,4 +1,5 @@
 import { DomainEvent, DomainEventProps } from '@shared/ddd';
+
 import { OrderItem } from '@modules/order/domain/orderItem';
 
 export class OrderCreated extends DomainEvent {
