+++ src/shared/application/context/AppRequestContext.ts
@@ -1,4 +1,5 @@
 import { RequestContext } from 'nestjs-request-context';
+
 // import { DatabaseTransactionConnection } from 'slonik';
 
 /**
