+++ src/shared/ddd/Entity.ts
@@ -1,4 +1,5 @@
 import { EventHandler } from '@shared/commons/core/Event.handler';
+
 import { UniqueEntityID } from './UniqueEntityID';
 
 export abstract class Entity<T> extends EventHandler {
