+++ test/app.e2e-spec.spec.ts
@@ -1,6 +1,7 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { INestApplication } from '@nestjs/common';
 import request from 'supertest';
+import { INestApplication } from '@nestjs/common';
+import { Test, TestingModule } from '@nestjs/testing';
+
 import { HttpModule } from '@modules/order/application/ms/http/http.module';
 
 describe('AppController (e2e)', () => {
